backend/requirements.txt
pydantic-settings
sqlmodel
ollama
FastAPI
uvicorn[standard]
jinja2
chromadb
langchain
langchain-community
langchain-text-splitters
pypdf
ragas